+++ src/Service/Ecr/src/EcrClient.php
@@ -65,6 +65,7 @@ class EcrClient extends AbstractApi
             case 'ap-southeast-1':
             case 'ap-southeast-2':
             case 'ap-southeast-3':
+            case 'ap-southeast-4':
             case 'ca-central-1':
             case 'eu-central-1':
             case 'eu-central-2':