+++ src/Service/CloudFormation/src/ValueObject/Stack.php
@@ -120,7 +120,7 @@ final class Stack
     private $rootId;
 
     /**
-     * Information about whether a stack's actual configuration differs, or has *drifted*, from it's expected configuration,
+     * Information about whether a stack's actual configuration differs, or has *drifted*, from its expected configuration,
      * as defined in the stack template and any values specified as template parameters. For more information, see Detecting
      * Unregulated Configuration Changes to Stacks and Resources.
      *