+++ src/Service/Lambda/CHANGELOG.md
@@ -7,6 +7,7 @@
 - AWS enhancement: Documentation updates.
 - AWS api-change: Release Lambda RuntimeManagementConfig, enabling customers to better manage runtime updates to their Lambda functions. This release adds two new APIs, GetRuntimeManagementConfig and PutRuntimeManagementConfig, as well as support on existing Create/Get/Update function APIs.
 - AWS api-change: Added `ap-southeast-4` region.
+- AWS api-change: Add Python 3.10 (python3.10) support to AWS Lambda
 
 ## 1.8.0