+++ src/Service/Lambda/CHANGELOG.md
@@ -8,6 +8,7 @@
 - AWS api-change: Release Lambda RuntimeManagementConfig, enabling customers to better manage runtime updates to their Lambda functions. This release adds two new APIs, GetRuntimeManagementConfig and PutRuntimeManagementConfig, as well as support on existing Create/Get/Update function APIs.
 - AWS api-change: Added `ap-southeast-4` region.
 - AWS api-change: Add Python 3.10 (python3.10) support to AWS Lambda
+- AWS api-change: Add Ruby 3.2 (ruby3.2) Runtime support to AWS Lambda.
 - AWS api-change: Add Java 17 (java17) support to AWS Lambda
 ## 1.8.0